Some small fixes.

+++ src/core/framework/network/steam/NGSteamPacketHandler.cpp
@@ -56,7 +56,7 @@ void NGSteamPacketHandler::readIncomingPacketsIntoQueue()
     
     bzero(packetMem, NW_MAX_PACKET_SIZE);
     
-    size_t packetSize = sizeof(packetMem);
+    uint32 packetSize = sizeof(packetMem);
     uint32_t incomingSize = 0;
     InputMemoryBitStream inputStream(packetMem, packetSize * 8);
     CSteamID fromId;
